# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paradox Connect API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/mess": {
      "get": {
        "tags": [
          "Mess"
        ],
        "summary": "List Messes",
        "operationId": "list_messes_mess_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Mess"
        ],
        "summary": "Create Mess",
        "operationId": "create_mess_mess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mess/{mess_id}/team": {
      "post": {
        "tags": [
          "Mess"
        ],
        "summary": "Assign Mess Team",
        "operationId": "assign_mess_team_mess__mess_id__team_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mess_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mess Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessAssignTeamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mess/{mess_id}/team/{team_user_id}/toggle_scan": {
      "put": {
        "tags": [
          "Mess"
        ],
        "summary": "Toggle Mess Scan",
        "operationId": "toggle_mess_scan_mess__mess_id__team__team_user_id__toggle_scan_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mess_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mess Id"
            }
          },
          {
            "name": "team_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team User Id"
            }
          },
          {
            "name": "logging",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Logging"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mess/allocate": {
      "post": {
        "tags": [
          "Mess"
        ],
        "summary": "Allocate Messes",
        "operationId": "allocate_messes_mess_allocate_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mess/my_mess": {
      "get": {
        "tags": [
          "Mess"
        ],
        "summary": "My Mess",
        "operationId": "my_mess_mess_my_mess_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mess/{mess_id}/scan": {
      "post": {
        "tags": [
          "Mess"
        ],
        "summary": "Scan Mess",
        "operationId": "scan_mess_mess__mess_id__scan_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mess_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mess Id"
            }
          },
          {
            "name": "slot",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slot"
            }
          },
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Day"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanQRRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mess/{mess_id}/statistics": {
      "get": {
        "tags": [
          "Mess"
        ],
        "summary": "Mess Statistics",
        "operationId": "mess_statistics_mess__mess_id__statistics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mess_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mess Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessAssignTeamRequest": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "MessAssignTeamRequest"
      },
      "MessCreateRequest": {
        "properties": {
          "mess_id": {
            "type": "string",
            "title": "Mess Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "preference": {
            "type": "string",
            "title": "Preference"
          }
        },
        "type": "object",
        "required": [
          "mess_id",
          "name",
          "capacity",
          "preference"
        ],
        "title": "MessCreateRequest"
      },
...
      "ScanQRRequest": {
        "properties": {
          "participant_id": {
            "type": "string",
            "title": "Participant Id"
          },
          "data": {
            "type": "string",
            "title": "Data"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "participant_id",
          "data",
          "timestamp"
        ],
        "title": "ScanQRRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}